requirements.txt
beautifulsoup4==4.9.1
Django==1.11.17
django-jsonfield==1.0.1
djangorestframework==3.11.1
python-http-client==3.2.7
requests==2.24.0
requests-oauthlib==1.3.0
urllib3==1.25.9

